custom-dynclass-loader
- we produce a `-hive` and a not hive version of the connector.  Due to how
we strip the packages, some Kafka Connect clusters can throw a NoClassDefFound
error when it cannot find hadoop/hive (e.g. when using a rest catalog). This is fine,
however the refleciton code we use only convers ClassPathNotFound to nulls.

This brings in DynClasses.java from iceberg (apache license) and adds the additional
catch for the NoClassDefFound linkage error