custom-dynclass-loader (#216)
* custom-dynclass-loader

- we produce a `-hive` and a not hive version of the connector.  Due to how
we strip the packages, some Kafka Connect clusters can throw a NoClassDefFound
error when it cannot find hadoop/hive (e.g. when using a rest catalog). This is fine,
however the refleciton code we use only covers ClassPathNotFound to nulls.  This fixes 
the non-hive version from throwing in some clusters.